+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 * Moved PetitParser Examples to a separate Git repository.
 * Add a `surroundedBy` helper that silently consumes input before and after another parser.
 * Deprecate `hasSideEffect` in `MapParser` by considering all callbacks to have side-effects, the benefit of the added complications is negligible.
+* Initial support for indentation based grammars.
 
 ## 4.4.0